Melhorando desempenho da página principal

+++ models/models.go
@@ -92,8 +92,8 @@ type ProcInfoResult struct {
 
 // GeneralTotals - contains the summary from all DadosJusBr data
 type GeneralTotals struct {
-	AgencyAmount             int64
-	MonthlyTotalsAmount      int64
+	AgencyAmount             int
+	MonthlyTotalsAmount      int
 	StartDate                time.Time
 	EndDate                  time.Time
 	RemunerationRecordsCount int
